mqtt-iot/1.12/mongogw/requirements.txt
paho-mqtt==1.3.1
pymongo==3.7.0
